refactor(config): move NerdBot domain config into the app module
The framework's DiscordBotConfig carried NerdBot-specific settings (suggestions, alpha/project, badges, fun and generator config). Move those fields, their config classes and NerdBotConfig itself into the app module so discord-framework holds only generic bot config. App call sites now read domain config through SkyBlockNerdsBot.config() rather than the framework getConfig() accessor.
